+++ src/commands/ssh.ts
@@ -11,7 +11,7 @@ You should have received a copy of the GNU General Public License along with @p0
 import { authenticate } from "../drivers/auth";
 import { guard } from "../drivers/firestore";
 import { sshOrScp } from "../plugins/ssh";
-import { SshCommandArgs, provisionRequest, requestToSsh } from "./shared";
+import { SshCommandArgs, provisionRequest, requestToSsh } from "./shared/ssh";
 import yargs from "yargs";
 
 export const sshCommand = (yargs: yargs.Argv) =>
