+++ src/plugins/aws/ssh.ts
@@ -8,22 +8,27 @@ This file is part of @p0security/cli
 
 You should have received a copy of the GNU General Public License along with @p0security/cli. If not, see <https://www.gnu.org/licenses/>.
 **/
-import { SshRequest } from "../../commands/shared";
-import { AwsSsh } from "./types";
+import { CliRequest, Request } from "../../types/request";
+import { AwsPermissionSpec, AwsSsh, AwsSshRequest } from "./types";
 
-export const awsRequestToSsh: (request: AwsSsh) => SshRequest = (request) => {
-  const { permission, generated } = request;
-  const { instanceId, accountId, region } = permission.spec;
-  const { idc, ssh, name } = generated;
-  const { linuxUserName } = ssh;
-  const common = { linuxUserName, accountId, region, id: instanceId };
-  return !idc
-    ? { ...common, role: name, type: "aws", access: "role" }
-    : {
-        ...common,
-        idc,
-        permissionSet: name,
-        type: "aws",
-        access: "idc",
-      };
+export const awsSshProvider = {
+  requestToSsh: (request: AwsSsh): AwsSshRequest => {
+    const { permission, generated } = request;
+    const { instanceId, accountId, region } = permission.spec;
+    const { idc, ssh, name } = generated;
+    const { linuxUserName } = ssh;
+    const common = { linuxUserName, accountId, region, id: instanceId };
+    return !idc
+      ? { ...common, role: name, type: "aws", access: "role" }
+      : {
+          ...common,
+          idc,
+          permissionSet: name,
+          type: "aws",
+          access: "idc",
+        };
+  },
+  toCliRequest: async (
+    request: Request<AwsPermissionSpec>
+  ): Promise<Request<CliRequest>> => ({ ...request, cliLocalData: undefined }),
 };
