+++ src/plugins/ssh/index.ts
@@ -11,11 +11,8 @@ You should have received a copy of the GNU General Public License along with @p0
 import {
   ScpCommandArgs,
   SshCommandArgs,
-  AwsSshRequest,
   SshRequest,
-  AwsSshRoleRequest,
-  AwsSshIdcRequest,
-} from "../../commands/shared";
+} from "../../commands/shared/ssh";
 import { PRIVATE_KEY_PATH } from "../../common/keys";
 import { print2 } from "../../drivers/stdio";
 import { Authn } from "../../types/identity";
@@ -23,7 +20,12 @@ import { assertNever, throwAssertNever } from "../../util";
 import { getAwsConfig } from "../aws/config";
 import { assumeRoleWithIdc } from "../aws/idc";
 import { ensureSsmInstall } from "../aws/ssm/install";
-import { AwsCredentials } from "../aws/types";
+import {
+  AwsCredentials,
+  AwsSshIdcRequest,
+  AwsSshRequest,
+  AwsSshRoleRequest,
+} from "../aws/types";
 import { assumeRoleWithOktaSaml } from "../okta/aws";
 import { withSshAgent } from "../ssh-agent";
 import {
